% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMFM.R
\name{CDMFM_new}
\alias{CDMFM_new}
\title{Collapsed sampler for MFM-SBM}
\usage{
CDMFM_new(
  data,
  data1,
  niterations,
  beta.a,
  beta.b,
  GAMMA,
  LAMBDA,
  initNClusters
)
}
\arguments{
\item{data1}{a n by n upper triangle for the adjacency matrix}

\item{niterations}{the total number of iterations in MFMSBM}

\item{beta.a}{hyperparameters for the prior on elements in Q matrix in Beta distribution}

\item{beta.b}{hyperparameters for the prior on elements in Q matrix in Beta distribution}

\item{GAMMA}{the parameter in Dirichlet distribution that controls the relative size of clusters}

\item{LAMBDA}{the parameter for Poisson distrition}

\item{initNClusters}{the initial number of clusters}

\item{datat}{a n by n adjancey matrix}
}
\value{
zout = clustering configuration, a n by 1 vector

Qout = probability matrix, a k by k matrix
}
\description{
Collapsed sampler for MFM-SBM
}
