% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFMSBM.R
\docType{data}
\name{dolphin_data}
\alias{dolphin_data}
\title{Dolphin Data Set}
\format{
An object of class \code{character} of length 5.
}
\source{
\url{https://networkrepository.com/soc-dolphins.php}
}
\usage{
dolphin_data
}
\description{
A social network of bottlenose dolphins. The dataset contains a list of all of links, where a link represents frequent associations between dolphins
diamonds.
}
\keyword{datasets}
