% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDahl.R
\name{getDahl}
\alias{getDahl}
\title{Dahl's method to summarize the samples from the MCMC}
\usage{
getDahl(MFMfit, burn)
}
\arguments{
\item{MFMfit}{the result from CDMFM_new function}

\item{burn}{the number of burn-in interations}
}
\value{
zout = estimated clustering configuration, a n by 1 vector

Qout = estimated probability matrix, a k by k matrix
}
\description{
Dahl's method to summarize the samples from the MCMC
}
