% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFMSBM.R
\name{loglike}
\alias{loglike}
\title{Log-likelihood function related to Jth observation}
\usage{
loglike(clusterassign, param, data, J, n)
}
\arguments{
\item{clusterassign}{n by 1 vector of cluster configuration}

\item{param}{k by k probability matrix}

\item{data}{n by n adjacency matrix}

\item{J}{observation index}

\item{n}{number of observations}
}
\value{
log-likelihood related to the jth observation
}
\description{
Log-likelihood function related to Jth observation
}
