% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFMSBM.R
\name{logmargs}
\alias{logmargs}
\title{Function for getting the collapsed sampler for MFM-SBM m(Aj)}
\usage{
logmargs(clusterassign, data, J, beta.a, beta.b)
}
\arguments{
\item{clusterassign}{n by 1 vector of clustering configuration}

\item{data}{n by n adjacency  matrix}

\item{J}{observation index}

\item{beta.a}{hyperparameters for the prior on elements in Q matrix in Beta distribution}

\item{beta.b}{hyperparameters for the prior on elements in Q matrix in Beta distribution}
}
\value{
m(Aj) i.e the collapsed sampler for MFM-SBM
}
\description{
logmargs function takes in five parameters: clusterassign, data, J, beta.a, beta.b, and returns the collapsed sampler for MFMSBM(i.e m(Aj))
}
