% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMFM.R
\name{CDMFM_new}
\alias{CDMFM_new}
\title{Collapsed sampler for MFM-SBM}
\usage{
CDMFM_new(
  data,
  data1,
  niterations,
  beta.a,
  beta.b,
  GAMMA,
  LAMBDA,
  initNClusters
)
}
\arguments{
\item{initNClusters}{}
}
\description{
Collapsed sampler for MFM-SBM
}
