% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFMSBM.R
\name{CDMFM_new}
\alias{CDMFM_new}
\title{Collapsed sampler for MFM-SBM}
\usage{
CDMFM_new(
  data,
  data1,
  niterations,
  beta.a,
  beta.b,
  GAMMA,
  LAMBDA,
  initNClusters
)
}
\arguments{
\item{data}{a n by n adjancey matrix}

\item{data1}{a n by n upper triangle for the adjacency matrix}

\item{niterations}{the total number of iterations in MFMSBM}

\item{beta.a}{hyperparameters for the prior on elements in Q matrix in Beta distribution}

\item{beta.b}{hyperparameters for the prior on elements in Q matrix in Beta distribution}

\item{GAMMA}{the parameter in Dirichlet distribution that controls the relative size of clusters}

\item{LAMBDA}{the parameter for Poisson distrition}

\item{initNClusters}{the initial number of clusters}
}
\value{
zout = clustering configuration, a n by 1 vector

Qout = probability matrix, a k by k matrix
}
\description{
CDMFM_new function returns a list with zout (clustering configuration, n by 1 vector), and Qout (probability matrix, k by k)
}
\examples{

n <- 10
A <- matrix(0,n,n)
AAA <- matrix(0,n,n)
CDMFM_new(data = A, data1 = AAA, niterations = 1, beta.a = 1, beta.b = 1, GAMMA=1, LAMBDA = 1, initNClusters = 9)

}
